--- component/ui/moti/components/image.tsx
@@ -1,5 +1,0 @@
-import { Image as RImage } from 'react-native'
-
-import { motify } from '../core'
-
-export const Image = motify(RImage)()
