--- component/ui/moti/components/index.ts
@@ -1,13 +1,0 @@
-export { Image } from './image'
-export { Image as MotiImage } from './image'
-export { SafeAreaView } from './safe-area-view'
-export { SafeAreaView as MotiSafeAreaView } from './safe-area-view'
-export { ScrollView } from './scroll-view'
-export { ScrollView as MotiScrollView } from './scroll-view'
-export { Text } from './text'
-export { Text as MotiText } from './text'
-export { View } from './view'
-export { View as MotiView } from './view'
-
-export { MotiProgressBar } from './progress'
-export type { MotiProgressBarProps } from './progress'
