--- component/ui/moti/components/safe-area-view.tsx
@@ -1,5 +1,0 @@
-import { SafeAreaView as RSafeAreaView } from 'react-native-safe-area-context'
-
-import { motify } from '../core'
-
-export const SafeAreaView = motify(RSafeAreaView)()
