--- component/ui/moti/components/scroll-view.tsx
@@ -1,5 +1,0 @@
-import { ScrollView as RScrollView } from 'react-native'
-
-import { motify } from '../core'
-
-export const ScrollView = motify(RScrollView)()
