--- component/ui/moti/components/text.tsx
@@ -1,5 +1,0 @@
-import { Text as RText } from 'react-native'
-
-import { motify } from '../core'
-
-export const Text = motify(RText)()
