--- component/ui/moti/components/view.tsx
@@ -1,5 +1,0 @@
-import { View as RView } from 'react-native'
-
-import { motify } from '../core'
-
-export const View = motify(RView)()
