--- component/ui/moti/core/constants/color-keys.ts
@@ -1,11 +1,0 @@
-export const colorKeys = [
-  'backgroundColor',
-  'borderBottomColor',
-  'borderColor',
-  'borderEndColor',
-  'borderLeftColor',
-  'borderRightColor',
-  'borderStartColor',
-  'borderTopColor',
-  'color',
-]
