--- component/ui/moti/core/constants/index.ts
@@ -1,2 +1,0 @@
-export * from './color-keys';
-export * from './package-name';
