--- component/ui/moti/core/constants/package-name.ts
@@ -1,2 +1,0 @@
-// import Package from '../../package.json'
-export const PackageName = 'moti'
