--- component/ui/moti/core/index.ts
@@ -1,8 +1,0 @@
-export { default as motify } from './motify'
-export { AnimatePresence } from 'framer-motion'
-
-export * from './types'
-export { default as useAnimationState } from './use-animator'
-export { default as useDynamicAnimation } from './use-dynamic-animation'
-export { useMotify } from './use-motify'
-export { colorKeys, PackageName } from './constants'
