--- component/ui/moti/index.tsx
@@ -1,2 +1,0 @@
-export * from './core'
-export * from './components'
