+++ packages/aargon-dropdown/src/index.ts
@@ -1,0 +1,3 @@
+export { AnimatedDropdown } from './AnimatedDropdown'
+export { DropdownItemComponent } from './DropdownItem'
+export * from './dropdownUtils'
